[Company][James]feature:implement should_throw_no_such_data_exception_when_delete_company_given_not_exist_id test

+++ src/main/java/com/thoughtworks/springbootemployee/service/CompanyService.java
@@ -55,8 +55,11 @@ public class CompanyService {
         return company;
     }
 
-    public void deleteById(int id) {
+    public void deleteById(int id) throws NoSuchDataException {
         //TODO
+        Optional<Company> companyOptional = companyRepository.findById(id);
+        if (!companyOptional.isPresent())
+            throw new NoSuchDataException("No such id company");
         companyRepository.deleteById(id);
 
     }
